txnprovider/txpool: move tests/txpool to txnprovider/txnpool/tests
