+++ txnprovider/txpool/tests/pool_test.go
@@ -14,11 +14,10 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
-package txpool
+package tests
 
 import (
 	"fmt"
-	"github.com/erigontech/erigon-lib/common/dir"
 	"math/big"
 	"testing"
 	"time"
@@ -26,13 +25,14 @@ import (
 	"github.com/holiman/uint256"
 	"github.com/stretchr/testify/require"
 
+	"github.com/erigontech/erigon-lib/common/dir"
 	"github.com/erigontech/erigon-lib/crypto"
 	"github.com/erigontech/erigon-lib/gointerfaces/sentryproto"
 	"github.com/erigontech/erigon-lib/log/v3"
 	"github.com/erigontech/erigon/execution/types"
 	"github.com/erigontech/erigon/rpc/requests"
-	"github.com/erigontech/erigon/tests/txpool/helper"
 	"github.com/erigontech/erigon/txnprovider/txpool"
+	"github.com/erigontech/erigon/txnprovider/txpool/tests/helper"
 )
 
 var (
